🐛Fix: 날씨 포스트 에러 수정
